@@ -1,41 +1,143 @@
 import math
-from typing import Dict, Sequence
+from typing import Dict, Tuple, Sequence
 
 import jax.numpy as jnp
 from jaxtyping import Array, Float
 from matplotlib.collections import LineCollection
 import matplotlib.pyplot as plt
 from matplotlib.ticker import MaxNLocator, NullFormatter
+import xarray as xr
 
-from ..trajectory import Batch, WHAT, Timeseries, TimeseriesEnsemble
+from ..timeseries import Set, Timeseries, TimeseriesEnsemble, WHAT
 from ..utils import UNIT
 from ..utils.unit import meters_to_kilometers, sq_meters_to_sq_kilometers, seconds_to_days
 
 
-class Evaluation(Batch):
+class Evaluation(Set):
+    """
+    Class that manages a dictionary of metric timeseries or timeseries ensemble and provides methods to access and
+    visualise them.
+
+    Attributes
+    ----------
+    _members : Dict[str, Timeseries | Sequence[TimeseriesEnsemble | Timeseries]]
+        A dictionary holding the metrics.
+    size : int
+        The number of metrics in the set.
+
+    Methods
+    -------
+    __init__(states: Dict[str, Timeseries | Sequence[TimeseriesEnsemble | Timeseries]]):
+        Initializes the Evaluation object with a dictionary of metric timeseries.
+
+    get(key: str) -> Timeseries | Sequence[TimeseriesEnsemble | Timeseries]:
+        Retrieves a metric by key.
+    
+    items() -> Tuple[str, Timeseries | Sequence[TimeseriesEnsemble | Timeseries]]:
+        Returns the items of the metrics dictionary.
+
+    keys() -> Tuple[str]:
+        Returns the keys of the metric timeseries dictionary.
+    
+    values() -> Tuple[Timeseries | Sequence[TimeseriesEnsemble | Timeseries]]:
+        Returns the values of the metrics dictionary.
+
+    plot(fig: plt.Figure, ti: int = None):
+        Plots the metric timeseries or timeseries ensemble up to a time index on the provided figure.
+
+    __getitem__(key: str) -> Timeseries | TimeseriesEnsemble:
+        Retrieves a metric by key.
+    """
+    
     _members: Dict[str, Timeseries | Sequence[TimeseriesEnsemble | Timeseries]]
     size: int
 
-    def __init__(
-            self,
-            states: Dict[str, Timeseries | Sequence[TimeseriesEnsemble | Timeseries]]
-    ):
+    def __init__(self, states: Dict[str, Timeseries | Sequence[TimeseriesEnsemble | Timeseries]]):
+        """
+        Initializes the Evaluation object with a dictionary of metric timeseries.
+
+        Parameters
+        ----------
+            states (Dict[str, Timeseries | Sequence[TimeseriesEnsemble | Timeseries]]): The initial metrics dictionary.
+        """
         self._members = states
         self.size = len(states)
 
     def get(self, key: str) -> Timeseries | Sequence[TimeseriesEnsemble | Timeseries]:
+        """
+        Retrieves a metric by key.
+
+        Parameters
+        ----------
+            key (str): The key of the metric to retrieve.
+
+        Returns
+        -------
+            Timeseries | Sequence[TimeseriesEnsemble | Timeseries]: The metrics corresponding to the key.
+        """
         return self._members.get(key)
 
-    def items(self) -> ((str, Timeseries | Sequence[TimeseriesEnsemble | Timeseries]), ...):
+    def items(self) -> Tuple[str, Timeseries | Sequence[TimeseriesEnsemble | Timeseries]]:
+        """
+        Returns the items of the metrics dictionary.
+
+        Returns
+        -------
+            Tuple[str, Timeseries | Sequence[TimeseriesEnsemble | Timeseries]]: The items of the metrics dictionary.
+        """
         return self._members.items()
 
-    def keys(self) -> (str, ...):
+    def keys(self) -> Tuple[str]:
+        """
+        Returns the keys of the metric timeseries dictionary.
+
+        Returns
+        -------
+            Tuple[str]: The keys of the metrics dictionary.
+        """
         return self._members.keys()
 
-    def values(self) -> (Timeseries | Sequence[TimeseriesEnsemble | Timeseries], ...):
+    def values(self) -> Tuple[Timeseries | Sequence[TimeseriesEnsemble | Timeseries]]:
+        """
+        Returns the values of the metrics dictionary.
+
+        Returns
+        -------
+            Tuple[Timeseries | Sequence[TimeseriesEnsemble | Timeseries]]: The values of the metrics dictionary.
+        """
         return self._members.values()
 
-    def plot(self, fig: plt.Axes, ti: int = None):
+    def to_dataarray(self) -> Dict[str, xr.DataArray]:
+        """
+        Converts the evaluation results to a dictionary of xarray DataArrays.
+
+        Returns
+        -------
+        Dict[str, xr.DataArray]
+            A dictionary where keys are the evaluation metric names and values are the corresponding xarray DataArrays.
+        """
+        return dict((key, value.to_dataarray()) for key, value in self.items())
+
+    def to_dataset(self) -> xr.Dataset:
+        """
+        Converts the evaluation results to an xarray Dataset.
+
+        Returns
+        -------
+        xr.Dataset
+            An xarray Dataset containing the evaluation results.
+        """
+        return xr.Dataset(self.to_dataarray())
+
+    def plot(self, fig: plt.Figure, ti: int = None):
+        """
+        Plots the metric timeseries or timeseries ensemble up to a time index on the provided figure.
+
+        Parameters
+        ----------
+            fig (plt.Figure): The figure to plot on.
+            ti (int, optional): The time index up to which to plot. If None, plots the full metric timeseries.
+        """
         if ti is None:
             ti = next(iter(self.values())).length
 
@@ -69,7 +171,7 @@ class Evaluation(Batch):
         add_legend: bool,
         ax: plt.Axes,
         has_label: bool = False
-    ) -> (Float[Array, ""], Float[Array, ""], Float[Array, "time"], str, str):
+    ) -> Tuple[Float[Array, ""], Float[Array, ""], Float[Array, "time"], str, str]:
         if isinstance(metric, Timeseries) or isinstance(metric, TimeseriesEnsemble):
             values, timedelta, unit = self.__parse_metric(metric)
 
@@ -107,7 +209,7 @@ class Evaluation(Batch):
         return min_values, max_values, timedelta, unit, what
 
     @staticmethod
-    def __guess_metrics_fig_layout(n_metrics: int) -> (int, int):
+    def __guess_metrics_fig_layout(n_metrics: int) -> Tuple[int, int]:
         min_n_cells = float("inf")
         best_layout = (0, 0)
         for n_rows in range(1, int(math.ceil(n_metrics ** 0.5)) + 1):
@@ -122,7 +224,7 @@ class Evaluation(Batch):
     @staticmethod
     def __parse_metric(
         metric: Timeseries | TimeseriesEnsemble
-    ) -> (Float[Array, "time"] | Float[Array, "member time"], Float[Array, "time"], str):
+    ) -> Tuple[Float[Array, "time"] | Float[Array, "member time"], Float[Array, "time"], str]:
         if isinstance(metric, Timeseries):
             values = metric.states[1:]
         else:
@@ -151,7 +253,7 @@ class Evaluation(Batch):
         timedelta_extended = jnp.tile(timedelta, (values.shape[0], 1))
         segments = jnp.concat([timedelta_extended[..., None], values[..., None]], axis=2)
         alpha = jnp.clip(1 / ((self.size / 10) ** (1 / 2)), .1, 1).item() / 2
-        lc = LineCollection(segments, alpha=alpha, color="black")  # noqa
+        lc = LineCollection(segments, alpha=alpha, color="black")
         ax.add_collection(lc)
 
     @staticmethod
@@ -191,4 +293,15 @@ class Evaluation(Batch):
         ax.set_ylabel(ylabel)
 
     def __getitem__(self, key: str) -> Timeseries | TimeseriesEnsemble:
+        """
+        Retrieves a metric by key.
+
+        Parameters
+        ----------
+            y (str): The key of the metric to retrieve.
+            
+        Returns
+        -------
+            Timeseries | TimeseriesEnsemble: The metric timeseries or timeseries ensemble corresponding to the key.
+        """
         return self._members[key]
