@@ -1,1 +1,1 @@
-from .version import __version__  # noqa
+from .version import __version__
