@@ -1,24 +1,39 @@
 import equinox as eqx
 
-from ..trajectory import Timeseries, TimeseriesEnsemble, Trajectory, TrajectoryEnsemble
+from ..timeseries import Timeseries, TimeseriesEnsemble, Trajectory, TrajectoryEnsemble
 from ..utils import WHAT
 from ._evaluation import Evaluation
 from ._evaluator import Evaluator
-from ._metric import METRIC
+from ._metric import METRIC_FUN
 
 
 class PairEvaluator(Evaluator):
-    @eqx.filter_jit
-    def __call__(
-            self,
-            reference_trajectory: Trajectory,
-            simulated_trajectory: Trajectory
-    ) -> [Timeseries, ...]:
-        print("Computing evaluation metrics...")
+    """
+    Class for evaluating a simulated trajectory using a set of predefined metrics.
 
+    Methods
+    -------
+    __call__(self, reference_trajectory: Trajectory, simulated_trajectory: Trajectory) -> Evaluation
+        Evaluates the `simulated_trajectory` against the `reference_trajectory` using the defined metrics.
+    """
+    
+    @eqx.filter_jit
+    def __call__(self, reference_trajectory: Trajectory, simulated_trajectory: Trajectory) -> Evaluation:
+        """
+        Evaluates the `simulated_trajectory` against the `reference_trajectory` using the defined metrics.
+        
+        Parameters
+        ----------
+            reference_trajectory (Trajectory): The reference trajectory to compare against.
+            simulated_trajectory (Trajectory): The simulated trajectory to be evaluated.
+        
+        Returns
+        -------
+            Evaluation: The result of the evaluation.
+        """
         metrics = {}
         for metric in self.metrics:
-            metric_name = METRIC[metric.metric]
+            metric_name = METRIC_FUN[metric.metric_fun]
 
             values = getattr(reference_trajectory, metric_name)(simulated_trajectory)
             metrics[metric_name] = Timeseries(
@@ -29,25 +44,40 @@ class PairEvaluator(Evaluator):
 
 
 class EnsembleEvaluator(Evaluator):
-    @eqx.filter_jit
-    def __call__(
-            self,
-            reference_trajectory: Trajectory,
-            simulated_trajectories: TrajectoryEnsemble
-    ) -> [Timeseries | TimeseriesEnsemble, ...]:
-        print("Computing evaluation metrics...")
+    """
+    Class for evaluating an ensemble of simulated trajectories using a set of predefined metrics.
 
+    Methods
+    -------
+    __call__(self, reference_trajectory: Trajectory, simulated_trajectory: TrajectoryEnsemble) -> Evaluation
+        Evaluates the `simulated_trajectories` ensemble against the `reference_trajectory` using the defined metrics.
+    """
+
+    @eqx.filter_jit
+    def __call__(self, reference_trajectory: Trajectory, simulated_trajectories: TrajectoryEnsemble) -> Evaluation:
+        """
+        Evaluates the `simulated_trajectories` ensemble against the `reference_trajectory` using the defined metrics.
+        
+        Parameters
+        ----------
+            reference_trajectory (Trajectory): The reference trajectory to compare against.
+            simulated_trajectory (Trajectory): The simulated ensemble of trajectories to be evaluated.
+        
+        Returns
+        -------
+            Evaluation: The result of the evaluation.
+        """
         metrics = {}
         for metric in self.metrics:
-            metric_name = METRIC[metric.metric]
+            metric_fun = METRIC_FUN[metric.metric_fun]
 
-            ensemble = getattr(simulated_trajectories, metric_name)(reference_trajectory)
+            ensemble = getattr(simulated_trajectories, metric_fun)(reference_trajectory)
             ensemble = TimeseriesEnsemble(
                 ensemble, reference_trajectory.times, metric.what, metric.unit
             )
 
             crps = simulated_trajectories.crps(
-                reference_trajectory, distance_func=getattr(Trajectory, metric_name)
+                reference_trajectory, distance_func=getattr(Trajectory, metric_fun)
             )
             crps = Timeseries(
                 crps, reference_trajectory.times, WHAT.crps, metric.unit
@@ -58,6 +88,6 @@ class EnsembleEvaluator(Evaluator):
                 mean, reference_trajectory.times, WHAT.mean, metric.unit
             )
 
-            metrics[metric_name] = (ensemble, crps, mean)
+            metrics[metric_fun] = (ensemble, crps, mean)
 
         return Evaluation(metrics)
