@@ -1,6 +1,6 @@
-from ._batch import Batch
+from ._set import Set
 from ._state import WHAT
-from .batch import StateBatch, TrajectoryBatch
 from .ensemble import TimeseriesEnsemble, TrajectoryEnsemble
+from .set import StateBatch, TrajectoryBatch
 from .state import Displacement, Location, State, Time
 from .timeseries import Timeseries, Trajectory
